backend/pyproject.toml
[project]
name = "carrier-sales-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2",
    "uvicorn[standard]>=0.38.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "argon2-cffi>=23.1.0",
    "pyjwt>=2.8.0",
]

[dependency-groups]
dev = [
    "ruff==0.14.5",
    "black==25.11.0",
    "isort==7.0.0",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["src"]
